iio: pressure: bm1390: Disallow allow oversampling
The BM1390 data-sheet states that the AVE_NUM field must have a fixed
value when IIR filter is in use. So, when IIR is used we must not allow
setting the oversampling.

The code is still there as it should be possible for an user to be able to
select whether to use IIR or simple averaging - averaging being faster.

Add #ifdefs to exclude the oversampling stuff for now.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>