tools: iio: iio_generic_buffer ensure alignment
The iio_generic_buffer can return garbage values when the total size of
scan data is not a multiple of the largest element in the scan. This can be
demonstrated by reading a scan, consisting, for example of one 4-byte and
one 2-byte element, where the 4-byte element is first in the buffer.

The IIO generic buffer code does not take into account the last two
padding bytes that are needed to ensure that the 4-byte data for next
scan is correctly aligned.

Add the padding bytes required to align the next sample with the scan size.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>
Fixes: e58537ccce73 ("staging: iio: update example application.")

---
Revision history
v5 => v6:
 - fix typo pointed by Andy (wan't => want).
v4 => v5:
 - drop unnecessary comment.
 - drop all other patches as those were already applied.
 - add Fixes-tag.
v3 => v4:
 - drop extra print and TODO coment
 - add comment clarifying alignment sizes