iio: test: Test GTS with uneven scales
The GTS helpers did flooring of scale when calculating available scales.
This resulted available-scales to be reported smaller than they should
when the division in scale computation resulted remainder greater than
half of the divider. (decimal part of result > 0.5)

Furthermore, when gains were computed based on scale, the gain resulting
from the scale computation was also floored. As a consequence the
floored scale reported by available scales did not match the gain
settings.

This was fixed by changing the available-scale computation and gain
computation both to do rounding instead of flooring.

Test scale computations using also values which do not result the
division to go even.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>