include: fix AM64/AM65x extended sections
Affected files: pru_cfg.h, pru_iep.h

In the linker.cmd files in the rest of the PSSP, the constants
registers (CREGISTER) names for the second parts of the IEP registers
and the CFG registers are called
PRU_IEPx_0x100
PRU_CFG_0x100

However, the registers were given a different name in the header files.
Update the header files to match the naming convention in the rest of
the PSSP files.

Reported by Angel Gavrailov

Signed-off-by: Nick Saulnier <nsaulnier@ti.com>