regmap: Add regval <=> real-val accessors
It is common for a few drivers to create table-based mapping from a physical
value <=> register value, and implement set and get functions to utilize
them. Typically these are used when a user requests driver to set some
human understandable value, and driver converts this to a register value and
writes it to hardware. Similarly, user requests for human understandable
values need reading a register and converting the register value to human
understandable value before returning it to user.

Add regmap helpers for doing this in order to reduce driver-specific
boilerplate.

Suggested-by: Marek Vasut <marek.vasut@gmail.com>
Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>