MAINTAINERS: add IIO GTS tests
Add undersigned as a maintainer for IIO GTS helper's KUnit tests.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>

---
This may or may not conflict with the filename fix here:
https://lore.kernel.org/all/CAO=gReFVhp7QK_XZRBO5vbv6fmFb4BdsZeQPSzWvuiz9UeQekA@mail.gmail.com/

Conflict should be very trivial so I didn't wait for the above fix to
get in IIO tree. I can rebase and resend if this appears to be a problem
though.