iio: Fix GTS Kunit test's config description
The short description line for the IIO GTS helper tests has been copied
from the other configs and is misleading.

Fix the short description line to indicate correct tests.

Fixes: cf996f039679 ("iio: test: test gain-time-scale helpers")
Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>

---
I think the cf996f039679 ("iio: test: test gain-time-scale helpers") is
not yet in any of the releases so maybe the fixes-tag is unnecessary.