regulator: bd71828: Support ROHM BD71851
The ROHM BD71851 is a power management IC which integrates 8 BUCKs and 4
LDOs. The PMIC has internal state-machine and it can support transitions
to RUN, SUSPEND and IDLE states. The LDOs 1 and 3 have two different
voltage range configurations that can be set at the manufacturing phase
by OTP. By default driver assumes low voltage ranges to be used because
the data-sheet indicates the higher voltage ranges to be an 'OTP option'.
The high voltage range can be indicated via device-tree property.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>

---
TODO:
- Consider still changing the default for LDO1 and 3 to be the 'high
  voltage range' to make it less likely the SW sets higher than intended
  voltages if wrong device-tree is applied.
- add dt-bindings