regulator: bd71828: Ensure unique struct names
The BD71828 driver has some global data structures which have very
generic names. This is confusing if driver supports more than one PMIC
models because names don't tell which PMIC the data is for. Furthermore,
if similar structures are needed by both models, then there will be a
naming conflict.

Prefix the names of global structures with the PMIC name to avoid naming
conflicts if different PMIC models are to be supported by the driver.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>