dt-bindings: mfd: ROHM BD71851 PMIC
The ROHM BD71851 is a power management IC which integrates 8 BUCKs and 4
LDOs. There is also an ADC and operation amplifier intended for current
/ temperature measurement and accumulation. RTC and 32.768 kHz clock
gate are also included. The PMIC can be customized via OTP and it has
options for operating as a main PMIC in multi-PMIC installation. Some
of the pins can also be used for GPO or GPI (including interrupt support).

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>