regmap: Allow setting IRQ domain name suffix
When multiple IRQ domains are created from the same device-tree node they
will get the same name based on the device-tree path. This will cause a
naming collision in debugFS when IRQ domain specific entries are created.

The regmap-IRQ creates per instance IRQ domains. This will lead to a
domain name conflict when a device which provides more than one
interrupt line uses the regmap-IRQ.

Add support for specifying an IRQ domain name suffix when creating a
regmap-IRQ controller.

The regmap-IRQ supports both the legacy IRQ domains and the linear IRQ
domains. New devices are not expected to be using the legacy domains so
support name suffixes only for linear domains and warn if suffix is
tried to be added for a legacy domain.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>
---
Revision history:
v2 => v3:
 - Drop name suffix support for the legacy domains