dt-bindings: regulator: ROHM BD72720
Add dt-binding doc for ROHM BD72720 regulators.

Every other property is already existing, except the
'ldon-head-millivolt'. It is used to set the BUCK10 mode to LDON_HEAD.

The LDON_HEAD is intended to be used when the BUCK10 is used to supply
the LDO1 .. LDO4. In this mode the BUCK10 voltage can't be controlled by
software, but the voltage is adjusted by PMIC to match the LDO1 .. LDO4
voltages with a given offset. Offset can be 50mV .. 300mV and is
changeable at 50mV steps.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>
---