dt-bindings: mfd: bd71815: Fix rsense and typos
The sense resistor used for measuring currents is typically some tens of
milli Ohms. It has accidentally been documented to be tens of mega Ohms.
Fix the size of this resistor and a few copy-paste errors while at it.

Drop the unsuitable 'rohm,charger-sense-resistor-ohms' property (which
can't represent resistors smaller than one Ohm), and introduce a new
'rohm,charger-sense-resistor-micro-ohms' property with appropriate
minimum, maximum and default values instead.

Fixes: 4238dc1e6490 ("dt_bindings: mfd: Add ROHM BD71815 PMIC")
Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>

---
The sense resistor property has not been used by upstream drivers, but
it is nevertheless good idea to keep it here for the sake of the
completenes. Furthermore, there may be a respin of the charger driver
RFC series coming as I am working on another PMIC variant with similar
charging block. Also, there is now an updstream dts with BD71828 PMIC,
which can be controlled with same downstream charger driver as BD71815.