dt-bindings: regulator: ROHM BD73800 regulators
Add bindings for the BUCKs and LDOs on ROHM BD73800. The PMIC state
specific voltages can be set in same fashion as with a few other ROHM
PMICs (for example with BD718[15,28,37,47,50,79]). Same properties are
recycled :)

The LDOs 1 and 4 can use different voltage ranges depending on the OTP
configuration.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>

---
Revision history:
v1 => v2:
 - Description of the state specific voltages moved to own file and
   referenced from it (because multiple PMIC bindings used them).

The dt checker's documentation reference checker warns about the link to
the MFD binding - which will be introduced in next commit. These commits
have kind of a chicken and egg issue as MFD is referencing regulator
properties introduced in this patch.

I could squash these but IMO, this is clearer for the reviewers. AFAICS,
Both MFD and regulator subsystems tend to be busy. I could also split
adding the MFD comment reference to own patch which would be after the
patch adding the MFD binding, but having own patch for just adding the
comment is a bit meh.

So, I decided to live with the broken doc-link between these two
patches.