dt-bindings: regulator: ROHM BD71851 regulators
Add bindings for the BUCKs and LDOs on ROHM BD71851. The PMIC state
specific voltages can be set in same fashion as with a few other ROHM
PMICs (for example with BD718[15,28,37,47,50,79]). Same properties are
recycled :)

The LDOs 1 and 4 can use different voltage ranges depending on the OTP
configuration.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>

---
Internal changelog:
 - fixes to pass validation