gpio: bd72720: Fix register offset
The GPIO offset as a offset of the Linux's gpiochip's base does not
directly map to register bits because not all of the pins are always
used as GPIO. The driver omits the pins which are not used for GPIO -
but the hardware registers are still there.

The pin_no however is the number of a physical pin, and it can be
directly mapped to the register value. Use this for masking the value.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>