dominion: Randomly select cards for dominion game
Randomly select cards for a game of dominion. The cards and prizes are
listed in a text file. Program separates cards in 3 categories; Cards
which cost less than 4, cards which cost exactly 4, and cards which cost
more than 4.

Program takes 3 numbers as parameters, First number telling how many
cards should be selected from each of the categories.

Currently there are cards for 3 sets, the basic game (dominion), 4.th
expansion (prosperity) and 10.th expansion (empires). Additionally this
contains couple of extra cards which are just shown as 'Extras'.

This reflects the game set I am usually playing with. Currently there is
no way (other than modifying the list of cards) to select which game
sets the cards are picked from. (This is on my TODO-list).

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>