diceadmin: convert text storage to MySQL database
Parsing the card info from a text file grows old quite fast. Let's takea
time trip back to the late 90's / early 2000 and create a MySQL database
with the card information. This allows developing a few new features to
the 'dice', or maybe going further with the retro-trip and writing some
php :)

Well, lets create a program which goes through the cards from the text
file and fills the information to MySQL database. Furthermore, let's
also ask user to fill some other card properties so we can enchance the
randomizer based on the user preferences. (Like weigh cards which do
nasty things (attack) or cards which provide multiple actions or include
defensive cards when attacks get in the deck or...)

This converter expects a few tables to be created in the database prior
running. I will see if I later add a mysql dump to format the tables.

Finally, this is an odd mixture of Finnish and English, so sorry folks!

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>